Number of threads now tied to number of available proxies
Note that if there were more unsuccessful threads last runtime than there are available proxies this time, the program will end up creating multiple threads with the same proxy, leading to rate limiting.

This can be fixed if we think of a way to shrink the number of threads to meet the current number. Will likely cause us to redo work, but worth it.

This would also imply the need to create a way to upscale the thread count.